fix: use java clamp instead of own clamp
