+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/monitor/AutocraftingMonitorScreen.java
@@ -27,7 +27,6 @@ import static com.refinedmods.refinedstorage.common.support.Sprites.ERROR;
 import static com.refinedmods.refinedstorage.common.support.Sprites.ICON_SIZE;
 import static com.refinedmods.refinedstorage.common.util.IdentifierUtil.createIdentifier;
 import static com.refinedmods.refinedstorage.common.util.IdentifierUtil.createTranslation;
-import static com.refinedmods.refinedstorage.common.util.MathUtil.darkenARGB;
 import static net.minecraft.client.renderer.RenderPipelines.GUI_TEXTURED;
 
 public class AutocraftingMonitorScreen extends AbstractBaseScreen<AbstractAutocraftingMonitorContainerMenu>
@@ -478,5 +477,18 @@ public class AutocraftingMonitorScreen extends AbstractBaseScreen<AbstractAutocr
             button.setY(getTaskButtonY(i));
             button.visible = isTaskButtonVisible(button.getY());
         }
+    }
+
+    public static int darkenARGB(final int argb, final double percentage) {
+        final int alpha = (argb >> 24) & 0xFF;
+        int red = (argb >> 16) & 0xFF;
+        int green = (argb >> 8) & 0xFF;
+        int blue = argb & 0xFF;
+
+        red = (int) Math.max(0, red * (1 - percentage));
+        green = (int) Math.max(0, green * (1 - percentage));
+        blue = (int) Math.max(0, blue * (1 - percentage));
+
+        return (alpha << 24) | (red << 16) | (green << 8) | blue;
     }
 }
