+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/widget/ScrollbarWidget.java
@@ -147,8 +147,7 @@ public class ScrollbarWidget extends AbstractWidget {
     @Override
     public boolean mouseScrolled(final double x, final double y, final double scrollX, final double scrollY) {
         if (enabled) {
-            // TODO: Clamp
-            final int scrollDirection = Math.max(Math.min(-(int) scrollY, 1), -1);
+            final int scrollDirection = Math.clamp(-(int) scrollY, -1, 1);
             if (smoothScrolling) {
                 startScrollAnimation(scrollDirection);
             } else {
@@ -185,8 +184,7 @@ public class ScrollbarWidget extends AbstractWidget {
     }
 
     public void setOffset(final double offset) {
-        // TODO: Clamp
-        this.offset = Math.min(Math.max(0, offset), maxOffset);
+        this.offset = Math.clamp(offset, 0, maxOffset);
         if (listener != null) {
             listener.accept(this.offset);
         }
